Moving custom controls under their own folder and namespace.
