+++ SSMSObjectExplorerMenu.csproj
@@ -71,17 +71,17 @@
     <Compile Include="AddUserDefinedParameter.Designer.cs">
       <DependentUpon>AddUserDefinedParameter.cs</DependentUpon>
     </Compile>
-    <Compile Include="ArgumentControl.cs">
+    <Compile Include="controls\ArgumentControl.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="ArgumentControl.Designer.cs">
+    <Compile Include="controls\ArgumentControl.Designer.cs">
       <DependentUpon>ArgumentControl.cs</DependentUpon>
     </Compile>
     <Compile Include="Constants.cs" />
-    <Compile Include="DefaultValueControl.cs">
+    <Compile Include="controls\DefaultValueControl.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="DefaultValueControl.Designer.cs">
+    <Compile Include="controls\DefaultValueControl.Designer.cs">
       <DependentUpon>DefaultValueControl.cs</DependentUpon>
     </Compile>
     <Compile Include="EnterUserDefinedArguments.cs">
@@ -97,7 +97,7 @@
     <Compile Include="objects\ComboBoxItem.cs" />
     <Compile Include="objects\CustomListDefaultValueModel.cs" />
     <Compile Include="objects\MenuItemErrorModel.cs" />
-    <Compile Include="objects\ObservableListView.cs">
+    <Compile Include="controls\ObservableListView.cs">
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="objects\StringListItem.cs" />
@@ -186,7 +186,7 @@
     <EmbeddedResource Include="AddUserDefinedParameter.resx">
       <DependentUpon>AddUserDefinedParameter.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="ArgumentControl.resx">
+    <EmbeddedResource Include="controls\ArgumentControl.resx">
       <DependentUpon>ArgumentControl.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="EnterUserDefinedArguments.resx">
@@ -206,6 +206,7 @@
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(VSToolsPath)\VSSDK\Microsoft.VsSDK.targets" Condition="'$(VSToolsPath)' != ''" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
